tests: Move writer_test.go
